refactor referee code into submodule
