+++ .gitignore
@@ -1,0 +1,4 @@
+.idea/
+*.swp
+.soccer.wbproj
+*.pyc
